requirements.txt
# Основные зависимости для Telegram Userbot
telethon>=1.40.0
python-dotenv>=1.0.0

# Обработка данных и анализ
pandas>=1.5.0
matplotlib>=3.6.0
numpy>=1.21.0

# Асинхронная работа с файлами
aiofiles>=0.8.0

# База данных
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Кэширование и очереди
redis>=4.0.0

# Поиск и индексация
elasticsearch>=8.0.0

# HTTP запросы
requests>=2.28.0

# MQTT для интеграций
asyncio-mqtt>=0.16.0

# Мониторинг
prometheus-client>=0.16.0

# Дополнительные утилиты
python-dateutil>=2.8.0
pytz>=2021.1 